Fill attachment AID
It's stupid, but mkvinfo doesn't tell us the AID. I'm now making the assumption that the attachment ID is in the order in which the attachments are returned by mkvinfo, starting the count from 1. Calling mkvmerge <file> --identify does give the AID but I don't want to have to make another call and parse all that.