Don't crash during clean-up if files don't exist yet
At some stages in the encoding process the files may not have been created yet. In that case we would previously crash again on a missing file when calling os.remove, which prevents the clean-up of other files that were deleted in code underneath it. Now we catch that before it happens.