Don't start processing until file write is complete
Previously we would start processing when the file had just started copying, resulting in the processing of an incomplete file. This resulted in only the first few seconds being processed successfully, and thus a broken video. We need to wait until the file is completely copied, so wait until close_write instead of create.