Add Python-based watch script to periodically rescan folder
The system's inotify function is unreliable. If a folder gets created and quickly after, an event gets fired within that folder, we may miss that event because inotify didn't add the watcher to the new folder yet.
This periodic watching is dirty, but practical. We just have to check for existence of the file when processing the file. But that's fine.