Probe longer for videos where streams start later
Apparently, by default, FFMPEG only probes the first 5 seconds to see which streams are available. And if there are subtitle streams or audio streams that don't have any audio/subtitles for the first 5 seconds, they won't be detected.
That's not usually a problem, especially since I don't convert subtitles yet. However it is a problem for this one video file I'm converting now which doesn't have any audio for the first 5 seconds during a title sequence.